requirements.txt
Django==4.0.3
django-import-export==2.7.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
fontawesomefree==6.1.1
matplotlib==3.4.3
pytz==2021.3
regex==2021.8.3
tablib==3.2.0
XlsxWriter==3.0.1
xlwt==1.3.0
Pillow==8.4.0